Moved to and created class packages for organization changed access modifiers accordingly so that values can be accessed (the specific modifiers can be changed from pub to private later just did public for quicker/easier implementation)
